fix(v9): 精度门两处 bug — norm weights dtype cast + gate 换 _microbatch_backward
Bug 1 (megatron_trainer.py:370): model.cuda() 不带 dtype cast，Apex 缺失时 torch
fallback RMSNorm 权重停在 fp32，与 bf16 激活 mismatch → loss 偏 0.28+。
改为 model.cuda().to(params_dtype)。

Bug 2 (v9_accuracy_gate.py:219): 精度门调 train_batch，后者在返回前执行
clip_grad_norm_ + optimizer.step，读出来的是截断后的梯度，与 TorchActor
的原始梯度不可比。改为 _microbatch_backward (forward+backward,无clip无step)
+ 显式 _finalize_model_grads()（TP=1/PP=1 下为 no-op，但对齐完整路径）。

+++ toy_rl/trainer/megatron_trainer.py
@@ -367,7 +367,11 @@ class MegatronTrainer:
             parallel_output=True,
             share_embeddings_and_output_weights=self.tie,
         )
-        self.model = model.cuda()
+        # 源的 Apex FusedRMSNorm 在 build 时就尊重 TransformerConfig.params_dtype；
+        # 本机 Apex 未装，mcore fallback 到 torch.nn.RMSNorm，后者初始化为 fp32
+        # 且不读 config.params_dtype。不显式 cast 会导致 norm 权重停在 fp32，
+        # 与 bf16 激活产生 dtype mismatch → 精度门 loss 偏离 0.28+。
+        self.model = model.cuda().to(params_dtype)
         self.model.train()
 
         # 源 model.py:539 把 mcore 的 finalize_model_grads 挂到 config 上，由
